pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "midnight-sdk"
version = "0.1.0"
description = "The first Python SDK for the Midnight blockchain"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["midnight", "blockchain", "zk", "zero-knowledge", "sdk"]
dependencies = [
    "httpx>=0.27",
    "websockets>=12",
    "pydantic>=2",
    "typer>=0.12",
    "rich>=13",
    "mnemonic>=0.21",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "ipython>=8.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "respx>=0.21",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.scripts]
midnight = "midnight_sdk.cli:cli_main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
